requirements.txt
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
moviepy>=1.0.3
python-dotenv>=1.0.0
pytz>=2023.3
